Refactor Repose controller to remove config and test calls
