+++ app/clients/IDockerClient.java
@@ -1,6 +1,7 @@
 package clients;
 
 import com.google.inject.ImplementedBy;
+import exceptions.InternalServerException;
 import models.Cluster;
 import models.Container;
 import models.User;
@@ -13,6 +14,10 @@ import java.util.List;
 @ImplementedBy(SpotifyDockerClient.class)
 public interface IDockerClient {
 
-    List<Container> getReposeContainers(Cluster cluster, User user);
+    List<Container> getReposeContainers(Cluster cluster, User user) throws InternalServerException;
+
+    boolean startReposeInstance(Cluster cluster, String containerId) throws InternalServerException;
+
+    boolean stopReposeInstance(Cluster cluster, String containerId) throws InternalServerException;
 
 }
