+++ conf/routes
@@ -17,11 +17,15 @@ GET        /app/auth                                    controllers.Login.index(
 
 # Reposes
 GET        /app/repose/list                             controllers.Repose.list()
-GET        /app/repose/:id/configurations               controllers.Repose.configurations(id)
-POST       /app/repose/:id/upload                       controllers.Repose.uploadReposeConfigs(id)
-POST       /app/repose/test/:id                         controllers.Repose.test(id)
 GET        /app/repose/stop/:id                         controllers.Repose.stop(id)
 GET        /app/repose/start/:id                        controllers.Repose.start(id)
+
+# Test
+POST       /app/test/:id                                controllers.Test.test(id)
+
+# Configuration
+GET        /app/configuration/:id/configurations        controllers.Configuration.configurations(id)
+POST       /app/configuration/:id/upload                controllers.Configuration.uploadReposeConfigs(id)
 
 # Map static resources from the /public folder to the /assets URL path
 GET        /assets/*file        controllers.Assets.versioned(path="/public", file: Asset)
